Move coverage tests into per-module test files
I moved the tests that previously lived in test_coverage_additions.py
into the per-module test files (test_config.py, test_labels.py, etc.)
that match the source modules they exercise. This matches the existing
test_<module>.py naming convention in the repo so future maintainers
can find coverage for a given module without grepping a catch-all file.

I also dropped a duplicate test_get_owners_and_repositories_handles_user_prefix
test in test_search.py that exactly matched the existing
test_get_owners_and_repositories_with_user test, and kept the
owner: prefix variant which was not previously covered.

All 191 tests still pass with 100% coverage on every source module.
