+++ test_search.py
@@ -1,9 +1,11 @@
 """Unit tests for the search module."""
 
 import unittest
-from unittest.mock import MagicMock
+from unittest.mock import MagicMock, PropertyMock, patch
 
-from search import get_owners_and_repositories, search_issues
+import github3
+import requests
+from search import get_owners_and_repositories, print_error_messages, search_issues
 
 
 class TestSearchIssues(unittest.TestCase):
@@ -150,3 +152,120 @@ class TestGetOwnerAndRepository(unittest.TestCase):
         result = get_owners_and_repositories("user:owner1")
         self.assertEqual(result[0].get("owner"), "owner1")
         self.assertIsNone(result[0].get("repository"))
+
+    def test_get_owners_and_repositories_handles_owner_prefix(self):
+        """Test get owner: prefix sets the owner."""
+        result = get_owners_and_repositories("owner:octocat")
+        self.assertEqual(result[0]["owner"], "octocat")
+
+
+class TestSearchCoverageGaps(unittest.TestCase):
+    """Covers search.py rate-limit, exception, and parser branches."""
+
+    @patch("search.sleep", return_value=None)
+    def test_wait_for_api_refresh_retries_then_succeeds(self, mock_sleep):
+        """Low rate limit sleeps and then continues once the limit refills."""
+        iterator = MagicMock()
+        # Sequence: low first, then refilled, then refilled (final exit).
+        type(iterator).ratelimit_remaining = PropertyMock(side_effect=[1, 30, 30])
+        iterator.__iter__.return_value = iter([MagicMock(title="I1")])
+
+        connection = MagicMock()
+        connection.search_issues.return_value = iterator
+
+        issues = search_issues(
+            "is:open", connection, [{"owner": "o", "repository": "r"}]
+        )
+        self.assertEqual(len(issues), 1)
+        mock_sleep.assert_called()  # We did sleep at least once.
+
+    @patch("search.sleep", return_value=None)
+    def test_wait_for_api_refresh_exceeds_max_retries(self, _mock_sleep):
+        """RuntimeError after the maximum number of rate-limit retries."""
+
+        iterator = MagicMock()
+        iterator.ratelimit_remaining = 0  # always too low
+        iterator.__iter__.return_value = iter([])
+
+        connection = MagicMock()
+        connection.search_issues.return_value = iterator
+
+        with self.assertRaises(RuntimeError):
+            search_issues("is:open", connection, [])
+
+    def test_periodic_refresh_after_full_page(self):
+        """Refresh is invoked after each full page of results."""
+
+        # 101 issues forces the modulo branch (idx % issues_per_page == 0).
+        page_size_threshold = 100
+        issues_list = [MagicMock(title=f"I{i}") for i in range(page_size_threshold + 1)]
+
+        iterator = MagicMock()
+        iterator.ratelimit_remaining = 30
+        iterator.__iter__.return_value = iter(issues_list)
+
+        connection = MagicMock()
+        connection.search_issues.return_value = iterator
+
+        # Rate-limit bypass keeps the refresh call a no-op so we can assert the
+        # full iteration completed without exercising the sleep path.
+        result = search_issues("is:open", connection, [], rate_limit_bypass=True)
+        self.assertEqual(len(result), page_size_threshold + 1)
+
+    def _assert_exception_exits(self, exception_instance):
+
+        iterator = MagicMock()
+        iterator.ratelimit_remaining = 30
+        iterator.__iter__.side_effect = exception_instance
+
+        connection = MagicMock()
+        connection.search_issues.return_value = iterator
+
+        with self.assertRaises(SystemExit):
+            search_issues(
+                "is:open",
+                connection,
+                [{"owner": "o", "repository": "r"}],
+                rate_limit_bypass=True,
+            )
+
+    def test_forbidden_error_exits(self):
+        """ForbiddenError from github3 triggers a clean SystemExit."""
+        resp = MagicMock(status_code=403)
+        resp.json.return_value = {"message": "forbidden"}
+        self._assert_exception_exits(github3.exceptions.ForbiddenError(resp))
+
+    def test_not_found_error_exits(self):
+        """NotFoundError from github3 triggers a clean SystemExit."""
+        resp = MagicMock(status_code=404)
+        resp.json.return_value = {"message": "not found"}
+        self._assert_exception_exits(github3.exceptions.NotFoundError(resp))
+
+    def test_connection_error_exits(self):
+        """ConnectionError wrapping a requests error triggers SystemExit."""
+        self._assert_exception_exits(
+            github3.exceptions.ConnectionError(requests.ConnectionError("boom"))
+        )
+
+    def test_authentication_failed_exits(self):
+        """AuthenticationFailed from github3 triggers a clean SystemExit."""
+        resp = MagicMock(status_code=401)
+        resp.json.return_value = {"message": "auth failed"}
+        self._assert_exception_exits(github3.exceptions.AuthenticationFailed(resp))
+
+    def test_unprocessable_entity_exits(self):
+        """UnprocessableEntity from github3 triggers a clean SystemExit."""
+        resp = MagicMock(status_code=422)
+        resp.json.return_value = {
+            "message": "Validation Failed",
+            "errors": [{"message": "bad query"}],
+        }
+        self._assert_exception_exits(github3.exceptions.UnprocessableEntity(resp))
+
+    def test_print_error_messages_with_errors_attr(self):
+        """print_error_messages iterates over .errors when present."""
+
+        error = MagicMock()
+        error.errors = [{"message": "bad query"}, {"message": "another"}]
+        # Should not raise.
+        print_error_messages(error)
